pyproject.toml
[project]
name = "django-myuser"
version = "0.1.0"
description = ""
authors = [
    {name = "Sanjaya Subedi"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "django (>=5.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt[crypto] (>=5.5.1,<6.0.0)",
    "django-allauth (>=65.11.1,<66.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "dj-rest-auth (>=7.0.1,<8.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
pytest-cov = "^6.2.1"

